+++ Mailer/extensions/mailer-module/src/Models/Sender/MailerRuntimeException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Remp\MailerModule\Models\Sender;
+
+use Exception;
+
+class MailerRuntimeException extends Exception
+{
+
+}